fix a test on 32 bits computer

+++ theano/tests/test_tutorial.py
@@ -578,7 +578,9 @@ class T_examples(unittest.TestCase):
 
     def test_examples_8(self):
         from theano import shared
-        state = shared(0)
+        # Force the dtype to int64 to work correctly on 32 bits computer.
+        # Otherwise, it create by default a int32 on 32 bits computer.
+        state = shared(numpy.int64(0))
         inc = T.iscalar('inc')
         accumulator = function([inc], state, updates=[(state, state+inc)])
